+++ components/molecules/HighlightInput/highlight-input-form.tsx
@@ -4,22 +4,31 @@ import { createHighlights } from "lib/hooks/createHighlights";
 
 import { ToastTrigger } from "lib/utils/toast-trigger";
 import { ChangeEvent, useEffect, useRef, useState } from "react";
-import PullRequestHighlightCard from "../PullRequestHighlightCard/pull-request-highlight-card";
 import { useSWRConfig } from "swr";
 import useSupabaseAuth from "lib/hooks/useSupabaseAuth";
+import GhOpenGraphImg from "../GhOpenGraphImg/gh-open-graph-img";
+import { generateApiPrUrl } from "lib/utils/github";
+import { fetchGithubPRInfo } from "lib/hooks/fetchGithubPRInfo";
 
 const HighlightInputForm = (): JSX.Element => {
   const { user } = useSupabaseAuth();
   const { mutate } = useSWRConfig();
   const [isDivFocused, setIsDivFocused] = useState(false);
+  const [loading, setLoading] = useState(false);
   const textAreaRef = useRef<HTMLTextAreaElement>(null);
   const [bodyText, setBodyText] = useState("");
   const [row, setRow] = useState(1);
   const [title, setTitle] = useState("");
+  const [charCount, setCharCount] = useState(0);
   const [pullrequestLink, setPullRequestLink] = useState("");
   const ref = useRef<HTMLFormElement>(null);
   let rowLomit = 5;
+  const charLimit = 500;
   let messageLastScrollHeight = textAreaRef.current ? textAreaRef.current?.scrollHeight : 50;
+
+  const validCharLimit = () => {
+    return charCount - pullrequestLink.length <= charLimit;
+  };
 
   useEffect(() => {
     const checkIfClickedOutside = (e: globalThis.MouseEvent) => {
@@ -61,31 +70,43 @@ const HighlightInputForm = (): JSX.Element => {
   const handlePostHighlight = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
 
-    // Trigger api call to post highlight
+    // Regex check for github pull request link match
     const pullLink = bodyText.match(/((https?:\/\/)?(www\.)?github\.com\/[^\/]+\/[^\/]+\/pull\/[0-9]+)/);
 
-    const link = pullLink && new URL(pullLink as unknown as string);
     const [url] = pullLink || [];
     const highlight = bodyText.replace(url as string, "");
 
-    if (url === null || url === undefined || url === "" || url.length === 0 || link?.hostname !== "github.com") {
-      ToastTrigger({ message: "A valid Pull request Link is required", type: "error" });
-      return;
-    } else {
-      const res = await createHighlights({
-        highlight,
-        title,
-        url: url || ""
-      });
 
-      if (res) {
-        mutate(`users/${user?.user_metadata.user_name}/highlights`);
-        setBodyText("");
-        setTitle("");
-        setIsDivFocused(false);
-        ToastTrigger({ message: "Highlight uploade success", type: "success" });
+    if (pullLink && url) {
+      const { apiPaths } = generateApiPrUrl(url);
+      const { repoName, orgName, issueId } = apiPaths;
+      setLoading(true);
+      // Api validation to check validity of github pull request link match
+      const res = await fetchGithubPRInfo(orgName, repoName, issueId);
+
+      if (res.isError) {
+        setLoading(false);
+        ToastTrigger({ message: "A valid Pull request Link is required", type: "error" });
+        return;
+
       } else {
-        ToastTrigger({ message: "An error occured!!!", type: "error" });
+        setLoading(true);
+        const res = await createHighlights({
+          highlight,
+          title,
+          url: url
+        });
+
+        setLoading(false);
+        if (res) {
+          mutate(`users/${user?.user_metadata.user_name}/highlights`);
+          setBodyText("");
+          setTitle("");
+          setIsDivFocused(false);
+          ToastTrigger({ message: "Highlight Posted!", type: "success" });
+        } else {
+          ToastTrigger({ message: "An error occured!!!", type: "error" });
+        }
       }
     }
   };
@@ -112,14 +133,27 @@ const HighlightInputForm = (): JSX.Element => {
           ref={textAreaRef}
           rows={row}
           value={bodyText}
-          onChange={(e) => handleTextAreaInputChange(e)}
+          onChange={(e) => {
+            handleTextAreaInputChange(e);
+            setCharCount(e.target.value.length);
+          }}
         />
+        {isDivFocused && (
+          <p className="text-xs pb-2 text-light-slate-9 flex justify-end gap-1">
+            <span className={`${!validCharLimit() && "text-red-600"}`}>
+              {!validCharLimit()
+                ? `-${charCount - pullrequestLink.length - charLimit}`
+                : charCount - pullrequestLink.length}
+            </span>{" "}
+            / <span>{charLimit}</span>
+          </p>
+        )}
       </div>
 
-      {pullrequestLink && isDivFocused && <PullRequestHighlightCard prLink={pullrequestLink} />}
+      {pullrequestLink && isDivFocused && <GhOpenGraphImg githubLink={pullrequestLink} />}
 
       {isDivFocused && (
-        <Button disabled={!bodyText} className="ml-auto" type="primary">
+        <Button loading={loading} disabled={!bodyText || !validCharLimit()} className="ml-auto" variant="primary">
           Post
         </Button>
       )}