+++ pages/[filterName]/insights/index.tsx
@@ -28,8 +28,8 @@ const InsightsHub: WithPageLayout = () => {
         if (!currentUser?.data?.session || onboarded === false) {
           await router.push("/");
         }
-      } catch(e: unknown) {
-        await router.push("/");
+      } catch (e: unknown) {
+        router.push("/");
       }
     }
 
@@ -48,7 +48,7 @@ const InsightsHub: WithPageLayout = () => {
             <Search placeholder="Search repositories" className="max-w-full" name={"query"} />
           </div>
           <Link href={"/hub/insights/new"}>
-            <Button type="primary">Add Insight Page</Button>
+            <Button variant="primary">Add Insight Page</Button>
           </Link>
         </div>
       </div>
@@ -59,21 +59,21 @@ const InsightsHub: WithPageLayout = () => {
           : isError
             ? "Error..."
             : insightsData.map((insight, index) => {
-              return (
-                <InsightRow  key={`insights_${insight.id}`} user={user} insight={insight} />
-              );
-            })
-        }
+              return <InsightRow key={`insights_${insight.id}`} user={user} insight={insight} />;
+            })}
       </section>
 
       <Link
         passHref
         href={"/hub/insights/new"}
-        className="w-full bg-light-slate-4 text-lg text-light-slate-11 py-5 md:py-8 lg:py-10 rounded-lg text-center border border-light-slate-7">
-          Create a new Insight Page
+        className="w-full bg-light-slate-4 text-lg text-light-slate-11 py-5 md:py-8 lg:py-10 rounded-lg text-center border border-light-slate-7"
+      >
+        Create a new Insight Page
       </Link>
     </div>
-  ) : <></>;
+  ) : (
+    <></>
+  );
 };
 
 InsightsHub.PageLayout = HubLayout;