+++ pages/start.tsx
@@ -5,10 +5,10 @@ import { User } from "@supabase/supabase-js";
 
 import CompletedIcon from "img/icons/completed-icon.svg";
 import GitHubAuthActiveIcon from "img/icons/github-auth-active-icon.svg";
-import ChooseRepoIcon from "img/icons/choose-repo-icon.svg";
-import ChooseRepoActiveIcon from "img/icons/choose-repo-active-icon.svg";
-import PATIcon from "img/icons/pat-icon.svg";
-import PATActiveIcon from "img/icons/pat-active-icon.svg";
+import ChooseTimezoneIcon from "img/icons/choose-timezone.svg";
+import ChooseTimezoneActiveIcon from "img/icons/choose-timezone-active.svg";
+import ChooseInterestsIcon from "img/icons/choose-interests.svg";
+import ChooseInterestsActiveIcon from "img/icons/choose-interests-active.svg";
 import HighlightIcon from "img/icons/highlight-icon.svg";
 import GitHubIcon from "img/icons/github-icon.svg";
 import AddIcon from "img/icons/add-icon.svg";
@@ -32,6 +32,14 @@ import { captureAnayltics } from "lib/utils/analytics";
 import { getAvatarByUsername } from "lib/utils/github";
 import useStore from "lib/store";
 import Avatar from "../components/atoms/Avatar/avatar";
+import getInterestOptions from "lib/utils/getInterestOptions";
+import LanguagePill from "components/atoms/LanguagePill/LanguagePill";
+import { updateUser } from "lib/hooks/update-user";
+import { ToastTrigger } from "lib/utils/toast-trigger";
+import Select from "components/atoms/Select/select";
+import SelectOption from "components/atoms/Select/select-option";
+import { timezones } from "lib/utils/timezones";
+import { authSession } from "lib/hooks/authSession";
 
 type handleLoginStep = () => void;
 
@@ -49,7 +57,7 @@ const LoginStep1: React.FC<LoginStep1Props> = ({ handleLoginStep, user }) => {
 
   useEffect(() => {
     if (onboarded) {
-      router.push("/hub/insights").then(() => {});
+      router.push(`/user/${user?.user_metadata.user_name}`);
     } else if (onboarded === false && user && providerToken) {
       handleLoginStep();
     }
@@ -97,7 +105,7 @@ const LoginStep1: React.FC<LoginStep1Props> = ({ handleLoginStep, user }) => {
           </div>
         </div>
         <div>
-          <Button onClick={handleGitHubAuth} type="primary" className="w-full mt-3 md:mt-0 h-10">
+          <Button onClick={handleGitHubAuth} variant="primary" className="w-full mt-3 md:mt-0 h-10">
             Authenticate <Icon IconImage={GitHubIcon} className="ml-2" />
           </Button>
         </div>
@@ -108,117 +116,62 @@ const LoginStep1: React.FC<LoginStep1Props> = ({ handleLoginStep, user }) => {
 
 interface LoginStep2Props {
   handleLoginStep: handleLoginStep;
-  setRepoList: Function;
+  handleUpdateInterests: (interests: string[]) => void;
 }
 
-const LoginStep2: React.FC<LoginStep2Props> = ({ handleLoginStep, setRepoList }) => {
-  const { providerToken } = useSupabaseAuth();
-  const [orgName, setOrgName] = useState("");
+const LoginStep2: React.FC<LoginStep2Props> = ({
+  handleLoginStep,
+  handleUpdateInterests: handleUpdateInterestsParent
+}) => {
+  const [selectedInterests, setSelectedInterests] = useState<string[]>([]);
+  const interestArray = getInterestOptions();
 
   captureAnayltics("User Onboarding", "onboardingStep2", "visited");
 
-  const handleSlugChange = (value: string) => {
-    setOrgName(value);
+  const handleSelectInterest = (interest: string) => {
+    if (selectedInterests.length > 0 && selectedInterests.includes(interest)) {
+      setSelectedInterests((prev) => prev.filter((item) => item !== interest));
+    } else {
+      setSelectedInterests((prev) => [...prev, interest]);
+    }
   };
 
-  const handleAddPAT = async () => {
-    try {
-      if (providerToken) {
-        const ocktokit = new Octokit({ auth: providerToken });
-
-        let repoList = [];
-        let response: { data: { id: number; full_name: string }[] };
-
-        if (orgName) {
-          response = await ocktokit.request("GET /orgs/{org}/repos", {
-            org: orgName,
-            sort: "updated",
-            direction: "desc",
-            // eslint-disable-next-line
-            per_page: 50
-          });
-        } else {
-          response = await ocktokit.request("GET /user/repos", {
-            visibility: "public",
-            sort: "updated",
-            direction: "desc",
-            // eslint-disable-next-line
-            per_page: 50
-          });
-        }
-
-        repoList = response.data.map((repo) => {
-          const [repoOwner, repoName] = repo.full_name.split("/");
-
-          return {
-            repoId: repo.id,
-            repoName,
-            repoOwner
-          };
-        });
-
-        setRepoList(repoList);
-      }
-
-      // If valid, go to next step
-      handleLoginStep();
-    } catch (e) {
-      // If invalid, display error
-      console.error(e);
-    }
+  const handleUpdateInterest = async () => {
+    handleUpdateInterestsParent(selectedInterests);
+    handleLoginStep();
   };
 
   return (
     <>
-      <div className="login-step flex flex-col h-full gap-6">
+      <div className="login-step flex flex-col h-full lg:gap-28">
         <div>
           <div className="flex items-center gap-2 mb-4">
-            <Icon className="lg:hidden" IconImage={PATActiveIcon} size={48} />
+            <Icon className="lg:hidden" IconImage={ChooseInterestsActiveIcon} size={48} />
             <Title className="!text-sm !text-light-orange-9">Step Two</Title>
           </div>
           <div className="gap-2 mb-4">
-            <Title className="!text-2xl">Provide your Organization</Title>
+            <Title className="!text-2xl">Choose your interests</Title>
           </div>
           <div className="mb-4 text-left ">
             <Text className="!text-sm">
-              In order to provide fresh, and insightful data, we‘ll use your authenticated GitHub access token to fetch
-              public GitHub data. Here‘s how we‘re going to use your token:
+              Take a moment to select your interests to help us provide personalized project recommendations. By doing
+              so, you&apos;ll find projects that match your skills and aspirations.
             </Text>
           </div>
-          <div className="flex flex-col gap-1">
-            <div className="flex gap-2 items-start">
-              <div className="mt-0.5 shrink-0">
-                <Icon IconImage={HighlightIcon} />
-              </div>
-              <div className="mt-0">
-                <Text className="!text-base !text-light-slate-12">Index insights from git data</Text>
-              </div>
-            </div>
-            <div className="flex gap-2 items-start">
-              <div className="mt-0.5 shrink-0">
-                <Icon IconImage={HighlightIcon} />
-              </div>
-              <div className="mt-0">
-                <Text className="!text-base !text-light-slate-12">
-                  Fetch basic GitHub information from Pull Requests and Issues
-                </Text>
-              </div>
-            </div>
+          <div className="flex gap-3 flex-wrap">
+            {interestArray.map((topic, index) => (
+              <LanguagePill
+                onClick={() => handleSelectInterest(topic)}
+                classNames={`${(selectedInterests || []).includes(topic) && "!bg-light-orange-10 !text-white"}`}
+                topic={topic}
+                key={index}
+              />
+            ))}
           </div>
         </div>
-        <div className="flex flex-col gap-6">
-          <div>
-            <TextInput
-              placeholder="Organization Slug"
-              descriptionText="Grab the slug from GitHub, eg: 'open-sauced'."
-              classNames="!rounded-md !py-0.5 child:h-8 child:text-sm"
-              handleChange={handleSlugChange}
-            />
-          </div>
-          <Button onClick={handleAddPAT} type="primary" className="w-full mt-3 md:mt-0 h-10">
-            Continue
-          </Button>
-        </div>
+        <Button onClick={handleUpdateInterest} variant="primary" className="w-full mt-3 md:mt-0 h-10">
+          Confirm Selections
+        </Button>
       </div>
     </>
   );
@@ -226,128 +179,75 @@ const LoginStep2: React.FC<LoginStep2Props> = ({ handleLoginStep, setRepoList })
 
 interface LoginStep3Props {
   handleLoginStep: handleLoginStep;
-  checkFollowed: {
-    isClickedFollowed: boolean;
-    setIsClickedFollowed: React.Dispatch<React.SetStateAction<boolean>>;
-  };
-  repoList: LoginRepoObjectInterface[];
+  interests: string[];
 }
 
-const LoginStep3: React.FC<LoginStep3Props> = ({ repoList, checkFollowed }) => {
+const LoginStep3: React.FC<LoginStep3Props> = ({ interests }) => {
   captureAnayltics("User Onboarding", "onboardingStep3", "visited");
   const store = useStore();
   const router = useRouter();
-  const { sessionToken } = useSupabaseAuth();
+  const { user, sessionToken } = useSupabaseAuth();
+  const [timezone, setTimezone] = useState("");
 
-  const [isFollowing, setIsFollowing] = useState<boolean[]>(repoList.map(() => false));
-  const following = isFollowing.filter((follow) => follow);
-
-  const handleSkipAddRepo = async () => {
+  const handleUpdateTimezone = async () => {
     try {
-      const selectedRepos = repoList.filter((_, index) => isFollowing[index]);
-      const repos = selectedRepos.map((repo) => ({ id: repo.repoId, fullName: `${repo.repoOwner}/${repo.repoName}`}));
-
-      await fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth/onboarding`, {
+      const data = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth/onboarding`, {
         method: "POST",
         headers: {
+          "Content-Type": "application/json",
           Authorization: `Bearer ${sessionToken}`
         },
-        body: JSON.stringify({ repos })
+        body: JSON.stringify({ interests, timezone })
       });
 
-      store.onboardUser();
+      if (data.ok) {
+        store.onboardUser();
+        router.push(`/user/${user?.user_metadata.user_name}`);
+      } else {
+        console.error("Error onboarding user");
+      }
     } catch (e) {
-      // handle error
-    }
-
-    await router.push("/hub/insights");
-  };
-
-  const handleFollowRepo = (index: number) => {
-    if (!isFollowing[index]) {
-      setIsFollowing((prevState) => {
-        const newState = [...prevState];
-        newState[index] = !newState[index];
-        return newState;
-      });
-      checkFollowed.setIsClickedFollowed(true);
-    } else {
-      setIsFollowing((prevState) => {
-        const newState = [...prevState];
-        newState[index] = !newState[index];
-        return newState;
-      });
+      console.error(e);
     }
   };
 
   return (
     <>
-      <div className="flex flex-col h-full gap-5">
+      <div className="flex flex-col h-full lg:gap-36">
         <div>
           <div className="flex items-center gap-2 mb-4">
-            <Icon className="lg:hidden" IconImage={ChooseRepoActiveIcon} size={48} />
+            <Icon className="lg:hidden" IconImage={ChooseTimezoneActiveIcon} size={48} />
             <Title className="!text-sm !text-light-orange-9">Step Three</Title>
           </div>
           <div className="gap-2 mb-4">
-            <Title className="!text-2xl">Follow some Repositories</Title>
+            <Title className="!text-2xl">What time is it?</Title>
           </div>
           <div className="mb-4 text-left ">
             <Text className="!text-sm">
-              We‘ll provide insights on the repos you choose to follow. You can follow up to 10 repos.
+              Provide your timezone to help companies discover you and collaborate on open source projects.
             </Text>
           </div>
-          <div className="relative">
-            <div className="absolute z-10 inset-x-0 top-[-1rem] h-[2rem] bg-white blur"></div>
-            <div className="absolute z-10 inset-x-0 bottom-[-1rem] h-[2rem] bg-white blur"></div>
-            <div className="max-h-72 overflow-y-auto py-4">
-              {repoList.map((repo, index) => {
-                return (
-                  <div
-                    key={index}
-                    className="flex justify-between w-full border-[1px] rounded-lg border-light-slate-6 p-2 mb-2 gap-2"
-                  >
-                    <div className="flex flex-col items-start shrink-1">
-                      <div className="flex items-center gap-1">
-                        <Avatar
-                          isCached={true}
-                          className="shrink-0 h-4 w-4 rounded overflow-hidden"
-                          size={40}
-                          avatarURL={getAvatarByUsername(repo.repoOwner)}
-                          isCircle={false}
-                        />
 
-                        <Text className="!text-sm ">{`${repo.repoOwner}/`}</Text>
-                      </div>
-                      <div className="w-full">
-                        <Text className="!text-base !text-light-slate-12 break-all">{`${repo.repoName}`}</Text>
-                      </div>
-                    </div>
-                    <div className="shrink-0">
-                      <Button
-                        disabled={following.length >= 10 && !isFollowing[index]}
-                        onClick={() => handleFollowRepo(index)}
-                        type={isFollowing[index] ? "outline" : "default"}
-                      >
-                        {isFollowing[index] ? (
-                          "Following"
-                        ) : (
-                          <>
-                            Follow <Icon IconImage={AddIcon} size={8} className="shrink-0 ml-2" />
-                          </>
-                        )}
-                      </Button>
-                    </div>
-                  </div>
-                );
-              })}
-            </div>
+          <div className="flex flex-col gap-2">
+            <label>Time zone*</label>
+            <Select value={timezone} onChange={(e) => setTimezone(e.target.value)} required>
+              <SelectOption value="">Select time zone</SelectOption>
+              {timezones.map((timezone, index) => (
+                <SelectOption key={index} value={timezone.value}>
+                  {timezone.text}
+                </SelectOption>
+              ))}
+            </Select>
           </div>
         </div>
-        <div onClick={handleSkipAddRepo} className="flex justify-center gap-2">
-          <Title className="!text-sm font-semibold !text-light-orange-9 cursor-pointer">
-            {following.length > 0 ? "Continue" : "Skip this step"}
-          </Title>
-        </div>
+        <Button
+          variant="primary"
+          onClick={handleUpdateTimezone}
+          className="w-full mt-3 md:mt-0 h-10"
+          disabled={!timezone}
+        >
+          Continue
+        </Button>
       </div>
     </>
   );
@@ -357,14 +257,11 @@ const Login: WithPageLayout = () => {
   type LoginSteps = number;
 
   const { user } = useSupabaseAuth();
-  const { repoList, setRepoList } = useLoginRepoList();
 
   const highlighted = "!text-light-slate-12";
 
   const [currentLoginStep, setCurrentLoginStep] = useState<LoginSteps>(1);
-  const [isClickedFollowed, setIsClickedFollowed] = useState<boolean>(false);
-
-  const checkFollowed = { isClickedFollowed, setIsClickedFollowed };
+  const [interests, setInterests] = useState<string[]>([]);
 
   const handleLoginStep = async () => {
     setCurrentLoginStep((prevStep) => prevStep + 1);
@@ -376,15 +273,7 @@ const Login: WithPageLayout = () => {
         <section className="w-full max-w-50%  p-6 lg:p-9">
           <div className="flex items-center gap-2 mb-6">
             <ProgressPie
-              percentage={
-                currentLoginStep === 1
-                  ? 0
-                  : currentLoginStep === 2
-                    ? 33
-                    : currentLoginStep === 3 && !isClickedFollowed
-                      ? 66
-                      : 100
-              }
+              percentage={currentLoginStep === 1 ? 0 : currentLoginStep === 2 ? 33 : currentLoginStep === 3 ? 66 : 100}
             />
             <Title className="!text-2xl !tracking-tight">Let‘s get started</Title>
           </div>
@@ -404,23 +293,29 @@ const Login: WithPageLayout = () => {
           </div>
           <div className="hidden lg:flex gap-2 items-center mb-8">
             <Icon
-              IconImage={currentLoginStep === 2 ? PATActiveIcon : currentLoginStep < 2 ? PATIcon : CompletedIcon}
+              IconImage={
+                currentLoginStep === 2
+                  ? ChooseInterestsActiveIcon
+                  : currentLoginStep < 2
+                    ? ChooseInterestsIcon
+                    : CompletedIcon
+              }
               size={48}
             />
             <Text
               disabled={currentLoginStep !== 2}
               className={`!text-[16px] !font-medium ${currentLoginStep === 2 && highlighted}`}
             >
-              Provide your Organization
+              Choose your interests
             </Text>
           </div>
           <div className="hidden lg:flex gap-2 items-center mb-8">
             <Icon
               IconImage={
-                currentLoginStep === 3 && !isClickedFollowed
-                  ? ChooseRepoActiveIcon
+                currentLoginStep === 3
+                  ? ChooseTimezoneActiveIcon
                   : currentLoginStep < 3
-                    ? ChooseRepoIcon
+                    ? ChooseTimezoneIcon
                     : CompletedIcon
               }
               size={48}
@@ -429,16 +324,19 @@ const Login: WithPageLayout = () => {
               disabled={currentLoginStep !== 3}
               className={`!text-[16px]  ${currentLoginStep === 3 && highlighted}`}
             >
-              Choose some repositories
+              What time is it?
             </Text>
           </div>
         </section>
         <section className="w-full lg:max-w-[50%] p-9 rounded-lg lg:rounded-r-lg bg-white">
           {currentLoginStep === 1 && <LoginStep1 handleLoginStep={handleLoginStep} user={user} />}
-          {currentLoginStep === 2 && <LoginStep2 handleLoginStep={handleLoginStep} setRepoList={setRepoList} />}
-          {currentLoginStep >= 3 && (
-            <LoginStep3 handleLoginStep={handleLoginStep} repoList={repoList} checkFollowed={checkFollowed} />
+          {currentLoginStep === 2 && (
+            <LoginStep2
+              handleLoginStep={handleLoginStep}
+              handleUpdateInterests={(interests) => setInterests(interests)}
+            />
           )}
+          {currentLoginStep >= 3 && <LoginStep3 handleLoginStep={handleLoginStep} interests={interests} />}
         </section>
       </>
     </Card>