+++ stories/atoms/button.stories.tsx
@@ -16,45 +16,32 @@ const storyConfig = {
 
 export default storyConfig;
 
-
 const ButtonTemplate: ComponentStory<typeof Button> = (args) => <Button {...args}>Button</Button>;
 
 export const ButtonTypes: ComponentStory<typeof Button> = (args) => (
   <div className="flex gap-2">
-    <Button type="default">
-      Default Button
-    </Button>
-    <Button type="primary">
-      Primary Button
-    </Button>
-    <Button type="outline">
-      Outline Button
-    </Button>
-    <Button type="link">
-      Link Button
-    </Button>
+    <Button variant="default">Default Button</Button>
+    <Button variant="primary">Primary Button</Button>
+    <Button variant="outline">Outline Button</Button>
+    <Button variant="link">Link Button</Button>
   </div>
 );
 
 export const Default = ButtonTemplate.bind({});
-Default.args = { type: "default" };
+Default.args = { variant: "default" };
 
 export const Primary = ButtonTemplate.bind({});
-Primary.args = { type: "primary" };
+Primary.args = { variant: "primary" };
 
 export const Outline = ButtonTemplate.bind({});
-Outline.args = { type: "outline" };
+Outline.args = { variant: "outline" };
 
 export const Link = ButtonTemplate.bind({});
-Link.args = { type: "link" };
+Link.args = { variant: "link" };
 
 export const HasIcons: ComponentStory<typeof Button> = (args) => (
   <div className="flex flex-wrap gap-2">
-    <Button type="default" icon={<Icon IconImage={ForkIcon.src}></Icon>}>
-      Icon Left
-    </Button>
-    <Button type="default" iconRight={<Icon IconImage={ForkIcon.src}></Icon>}>
-      Icon Right
-    </Button>
+    <Button variant="default">Icon Left</Button>
+    <Button variant="default">Icon Right</Button>
   </div>
 );